1) closing sourceforge bug #1087788. rearranged the midi variables which had the channel number appended to their names into arrays, and replaced the related repetitive code with loops.
2)
The interface of the registry class used to require we
specify the type and size of the each value we read or
write. I've changed it to use function overloading for types
like bool, all variants of int, and std::string, and to use
a template for other types so they just use REG_BINARY
(which means any type). So we can simply pass the variable
to load/save and its type and size are implicit ; it's less
error prone and looks nicer.

It works correctly, but i've noticed problems loading my
existing registry settings because some of the types used to
be REG_BINARY for values that should have been REG_DWORD
(and iirc vice-versa), and the registry class complains
about type missmatch. To work around it i've changed the
path of the "root" key to software/psycle/psycledelics-1.7
(it's actually determined from the values in
version.hpp).... the drawback is that people "loose" their
old settings.


git-svn-id: http://svn.code.sf.net/p/psycle/code/trunk@1169 f55022e6-6845-49fe-a7ed-519222bdd1fd